# Day26/Ecommerce/docs/openapi.yaml
openapi: 3.0.0
info:
  title: ECommerce API
  version: 1.0.0
  description: Complete ecommerce platform with authentication, products, cart, orders, and order management
  contact:
    name: API Support
    email: support@ecommerce.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Order:
      type: object
      properties:
        id:
          type: string
          example: "d9b994f7-99ac-4c19-bd54-c14b60c8973a"
        user_id:
          type: integer
          example: 1
        total_amount:
          type: number
          format: float
          example: 1999.98
        status:
          type: string
          enum: [pending, shipped, delivered, cancelled]
          example: pending
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    OrderItem:
      type: object
      properties:
        id:
          type: string
          example: "e8a765f8-99bd-5d20-bf65-e25b71d8974b"
        order_id:
          type: string
          example: "d9b994f7-99ac-4c19-bd54-c14b60c8973a"
        product_id:
          type: integer
          example: 1
        quantity:
          type: integer
          example: 2
        price:
          type: number
          format: float
          example: 999.99
        created_at:
          type: string
          format: date-time
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, shipped, delivered, cancelled]
          example: shipped

    # OrderItems Schemas (NEW)
    OrderItemWithDetails:
      type: object
      properties:
        id:
          type: string
          example: "e8a765f8-99bd-5d20-bf65-e25b71d8974b"
        order_id:
          type: string
          example: "d9b994f7-99ac-4c19-bd54-c14b60c8973a"
        product_id:
          type: integer
          example: 1
        quantity:
          type: integer
          example: 2
        price:
          type: number
          format: float
          example: 999.99
        created_at:
          type: string
          format: date-time
        order_status:
          type: string
          example: pending
        user_email:
          type: string
          example: john@example.com
        product_name:
          type: string
          example: iPhone 15 Pro
    OrderWithItems:
      type: object
      properties:
        order:
          $ref: '#/components/schemas/Order'
        items:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                example: "e8a765f8-99bd-5d20-bf65-e25b71d8974b"
              order_id:
                type: string
                example: "d9b994f7-99ac-4c19-bd54-c14b60c8973a"
              product_id:
                type: integer
                example: 1
              quantity:
                type: integer
                example: 2
              price:
                type: number
                format: float
                example: 999.99
              product_name:
                type: string
                example: iPhone 15 Pro

    # Response Schemas
...
    OrderCreateResponse:
      type: object
      properties:
        message:
          type: string
          example: Order placed successfully
        order:
          $ref: '#/components/schemas/Order'
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
    OrdersResponse:
      type: array
      items:
        $ref: '#/components/schemas/Order'
    OrderStatusResponse:
      type: object
      properties:
        message:
          type: string
          example: Order status updated
        order:
          type: object
          properties:
            id:
              type: string
              example: "d9b994f7-99ac-4c19-bd54-c14b60c8973a"
            status:
              type: string
              example: shipped
            updated_at:
              type: string
              format: date-time
    OrderDeleteResponse:
      type: object
      properties:
        message:
          type: string
          example: Order deleted successfully

    # OrderItems Response Schemas (NEW)
    OrderItemsResponse:
      type: array
      items:
        $ref: '#/components/schemas/OrderItemWithDetails'
    OrderItemsDeleteResponse:
      type: object
      properties:
        message:
          type: string
          example: Order item deleted successfully
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Error description
        error:
          type: string
          example: Detailed error message
...
paths:
...
  /orders:
    post:
      tags: [Orders]
      summary: Create order from cart
      description: Create an order from cart items. Clears the user's cart after successful order creation.
      security:
        - BearerAuth: []
      responses:
        201:
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderCreateResponse'
              example:
                message: Order placed successfully
                order:
                  id: "d9b994f7-99ac-4c19-bd54-c14b60c8973a"
                  user_id: 1
                  total_amount: 1999.98
                  status: pending
                  created_at: "2023-10-26T10:30:00.000Z"
                  updated_at: "2023-10-26T10:30:00.000Z"
                items:
                  - order_id: "d9b994f7-99ac-4c19-bd54-c14b60c8973a"
                    product_id: 1
                    quantity: 2
                    price: 999.99
                  - order_id: "d9b994f7-99ac-4c19-bd54-c14b60c8973a"
                    product_id: 2
                    quantity: 1
                    price: 899.99
        400:
          description: Cart is empty
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: Cart is empty
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Failed to place order

    get:
      tags: [Orders]
      summary: Get user orders
      description: Retrieve all orders for the authenticated user.
      security:
        - BearerAuth: []
      responses:
        200:
          description: Orders retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrdersResponse'
              example:
                - id: "d9b994f7-99ac-4c19-bd54-c14b60c8973a"
                  user_id: 1
                  total_amount: 1999.98
                  status: pending
                  created_at: "2023-10-26T10:30:00.000Z"
                  updated_at: "2023-10-26T10:30:00.000Z"
                - id: "a8c765e8-88bd-4e12-af63-d25a98765432"
                  user_id: 1
                  total_amount: 899.99
                  status: delivered
                  created_at: "2023-10-25T14:20:00.000Z"
                  updated_at: "2023-10-26T09:15:00.000Z"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Failed to fetch orders
  /orders/{id}/status:
    patch:
      tags: [Orders]
      summary: Update order status
      description: Update order status. **Admin only**.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Order ID (UUID format)
          example: "d9b994f7-99ac-4c19-bd54-c14b60c8973a"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderStatusUpdate'
            example:
              status: shipped
      responses:
        200:
          description: Order status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderStatusResponse'
              example:
                message: Order status updated
                order:
                  id: "d9b994f7-99ac-4c19-bd54-c14b60c8973a"
                  status: shipped
                  updated_at: "2023-10-26T11:45:00.000Z"
        400:
          description: Invalid status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: Invalid status
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        403:
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        404:
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: Order not found
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Failed to update order status
  /orders/{id}:
    delete:
      tags: [Orders]
      summary: Delete order
      description: Delete an order and its items. **Admin only**.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Order ID (UUID format)
          example: "d9b994f7-99ac-4c19-bd54-c14b60c8973a"
      responses:
        200:
          description: Order deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDeleteResponse'
              example:
                message: Order deleted successfully
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        403:
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        404:
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: Order not found
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Failed to delete order

  # ==================== ORDER ITEMS ROUTES (NEW) ====================
  /order-items:
    get:
      tags: [OrderItems]
      summary: Get all order items
      description: |
        **Role-based access:**
        - **Admin**: Gets all order items with order & user details
        - **User**: Gets only their own order items
      security:
        - BearerAuth: []
      responses:
        200:
          description: Order items retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderItemsResponse'
              examples:
                adminResponse:
                  summary: Admin view - All order items
                  value:
                    - id: "e8a765f8-99bd-5d20-bf65-e25b71d8974b"
                      order_id: "d9b994f7-99ac-4c19-bd54-c14b60c8973a"
                      product_id: 1
                      quantity: 2
                      price: 999.99
                      created_at: "2023-10-26T10:30:00.000Z"
                      order_status: pending
                      user_email: john@example.com
                      product_name: iPhone 15 Pro
                    - id: "f9b876g9-10ce-6e31-cg76-f36c82e9085c"
                      order_id: "a8c765e8-88bd-4e12-af63-d25a98765432"
                      product_id: 2
                      quantity: 1
                      price: 899.99
                      created_at: "2023-10-25T14:20:00.000Z"
                      order_status: delivered
                      user_email: jane@example.com
                      product_name: Samsung Galaxy S23
                userResponse:
                  summary: User view - Own order items
                  value:
                    - id: "e8a765f8-99bd-5d20-bf65-e25b71d8974b"
                      order_id: "d9b994f7-99ac-4c19-bd54-c14b60c8973a"
                      product_id: 1
                      quantity: 2
                      price: 999.99
                      created_at: "2023-10-26T10:30:00.000Z"
                      order_status: pending
                      product_name: iPhone 15 Pro
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Failed to fetch order items
  /orders/{orderId}/items:
    get:
      tags: [OrderItems]
      summary: Get order items by order ID
      description: |
        Get items for a specific order.
        - **Admin**: Can view any order
        - **User**: Can only view their own orders
      security:
        - BearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
          description: Order ID (UUID format)
          example: "d9b994f7-99ac-4c19-bd54-c14b60c8973a"
      responses:
        200:
          description: Order items retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderWithItems'
              example:
                order:
                  id: "d9b994f7-99ac-4c19-bd54-c14b60c8973a"
                  user_id: 1
                  total_amount: 1999.98
                  status: pending
                  created_at: "2023-10-26T10:30:00.000Z"
                  updated_at: "2023-10-26T10:30:00.000Z"
                items:
                  - id: "e8a765f8-99bd-5d20-bf65-e25b71d8974b"
                    order_id: "d9b994f7-99ac-4c19-bd54-c14b60c8973a"
                    product_id: 1
                    quantity: 2
                    price: 999.99
                    product_name: iPhone 15 Pro
                  - id: "f9b876g9-10ce-6e31-cg76-f36c82e9085c"
                    order_id: "d9b994f7-99ac-4c19-bd54-c14b60c8973a"
                    product_id: 2
                    quantity: 1
                    price: 899.99
                    product_name: Samsung Galaxy S23
        403:
          description: Forbidden - Cannot access this order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: Forbidden
        404:
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: Order not found
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Failed to fetch order items
  /order-items/{id}:
    delete:
      tags: [OrderItems]
      summary: Delete order item
      description: Delete a specific order item. **Admin only**.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Order item ID (UUID format)
          example: "e8a765f8-99bd-5d20-bf65-e25b71d8974b"
      responses:
        200:
          description: Order item deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderItemsDeleteResponse'
              example:
                message: Order item deleted successfully
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        403:
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        404:
          description: Order item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: Order item not found
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Failed to delete order item

  # ==================== GOOGLE OAUTH ROUTES ====================
...